third

+++ src/main/java/com/grayb/maintanencerequestserver/Domain/maintenanceRequest/models/MaintenanceRequest.java
@@ -26,7 +26,7 @@ public class MaintenanceRequest {
 
     private String email;
 
-    private int aptNumber;
+    private Long aptNumber;
 
     private String description;
